refactor: move ConnectClient to the new domain
