+++ src/Framework/Http/ConnectServer/Client/Exceptions/RequestException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Give\ConnectClient\Exceptions;
+namespace Give\Framework\Http\ConnectServer\Client\Exceptions;
 
 use Give\Framework\Exceptions\Primitives\RuntimeException;
